requirements.txt
captum==0.6.0
efficientnet-pytorch==0.7.1
matplotlib
numpy
opencv-python-headless
pandas
Pillow
pycocotools
PyYAML
scikit-image
scikit-learn
scipy
seaborn
termcolor
tensorflow==2.0.0
timm==0.9.2
torch==1.12.1
torchvision==0.13.1
tqdm
xmltodict
zennit==0.5.1
